Points system added, Score displayed, UI and comments improved

+++ Team 5 Game/core/src/main/java/io/github/some_example_name/Points.java
@@ -1,5 +1,14 @@
 package io.github.some_example_name;
 
+/* This class is responsible for calculating the player score in the game.
+ * Points are earned by in game events but can also be lost due to 
+ * penalties such as being caught by the Dean.
+ * 
+ * If the player escapes, the remaining time left is multipled by 50 and 
+ * added to the score.
+ * 
+ * The final score is displayed at the end of the game.
+ */
 public class Points {
     private int baseScore;
     private int penalties;
@@ -23,7 +32,8 @@ public class Points {
         penalties += 200;
     }
 
-    public void calcPoints(float remainingTime){
+    // Called at the end of the game
+    public void calcPoints(double remainingTime){
         int timeBonus = (int) (remainingTime * 50);
         addPoints(timeBonus);
         finalScore = baseScore - penalties;
@@ -33,6 +43,7 @@ public class Points {
 
     }
 
+    // Getter for points
     public int getScore(){
         return finalScore;
     }
